fix: ensure table locations are normalized for S3DynamoDbLogStore
The URLs used when writing entries into the table should always be
normalized to ensure that the value of table_url() is identical to the
one that is being written into the DynamoDB table.

There may be some more testing required with the Delta/Spark
implementation of S3DynamoDbLogStore to ensure that they are always generating
URLs to the table with trailing slashes. Delta/Spark stores paths as
[org.apache.hadoop.fs.Path](https://hadoop.apache.org/docs/stable/api/org/apache/hadoop/fs/Path.html)
which should have URL-like semantics. Without treating trailing slashes
as required, this could lead to inconsistencies between the two
implementations.

Signed-off-by: R. Tyler Croy <rtyler@brokenco.de>